# project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UCRS API Documentation",
        "description": "University Course Registration System - Complete API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/submissions": {
            "get": {
                "tags": [
                    "Submissions"
                ],
                "summary": "List all submissions",
                "description": "Get paginated list of all submissions (Instructor/Admin)",
                "operationId": "b74a0b2c1b1fcd8f54f08af2f9811bd4",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submissions retrieved"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/submissions/{id}": {
            "get": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Get submission details",
                "description": "Get specific submission with details (Instructor/Admin)",
                "operationId": "cd2a59547a628d683933a63fb3b892cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission details"
                    },
                    "404": {
                        "description": "Submission not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Update submission",
                "description": "Update submission content (Student, before grading)",
                "operationId": "22ac9a2cfbd527dac44311d7f7fd5b2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content_url"
                                ],
                                "properties": {
                                    "content_url": {
                                        "type": "string",
                                        "format": "url"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Submission updated"
                    },
                    "403": {
                        "description": "Cannot update graded submission"
                    },
                    "404": {
                        "description": "Submission not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/assignments/{assignmentId}/submit": {
            "post": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Submit assignment",
                "description": "Create new submission for assignment (Student)",
                "operationId": "6608c408ff109cf6c6d4a0e818aaf533",
                "parameters": [
                    {
                        "name": "assignmentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content_url"
                                ],
                                "properties": {
                                    "content_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/submission.pdf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Submission created"
                    },
                    "422": {
                        "description": "Already submitted or validation error"
                    },
                    "404": {
                        "description": "Assignment not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/submissions/my-submissions": {
            "get": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Get my submissions",
                "description": "Get all submissions for current student",
                "operationId": "61f13ac636c8fad6f191356c3c7a1f94",
                "responses": {
                    "200": {
                        "description": "My submissions retrieved"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/submissions/{id}/grade": {
            "put": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Grade submission",
                "description": "Assign grade to submission (Instructor/Admin)",
                "operationId": "17476d9feec009fcefc4a06d2390969f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "score"
                                ],
                                "properties": {
                                    "score": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 95
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Submission graded"
                    },
                    "404": {
                        "description": "Submission not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}